pyproject.toml
[tool.poetry]
name = "finance-tracker"
version = "0.1.0"
description = ""
authors = ["Sudarshan Srinivasan <sudarshan85@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"